requirements.txt
Jinja2==2.7.3
MarkupSafe==0.23
PyYAML==3.11
Pygments==2.0.1
Unidecode==0.04.16
Werkzeug==0.9.6
blinker==1.3
click==3.3
dateparser==0.1.0
docutils==0.12
feedgenerator==1.7
gnureadline==6.3.3
markdown2==2.3.0
pelican==3.5.0
python-dateutil==2.2
python-frontmatter==0.1.4
pytz==2014.10
six==1.8.0
wheel==0.23.0
wsgiref==0.1.2
